+++ layers/Schema/packages/users/src/ConditionalOnComponent/CustomPosts/ModuleProcessors/FormInputs/FilterInputModuleProcessor.php
@@ -16,11 +16,13 @@ class FilterInputModuleProcessor extends AbstractFormInputModuleProcessor implem
     use DataloadQueryArgsSchemaFilterInputModuleProcessorTrait;
 
     public const MODULE_FILTERINPUT_AUTHOR_IDS = 'filterinput-author-ids';
+    public const MODULE_FILTERINPUT_AUTHOR_SLUG = 'filterinput-author-slug';
 
     public function getModulesToProcess(): array
     {
         return array(
             [self::class, self::MODULE_FILTERINPUT_AUTHOR_IDS],
+            [self::class, self::MODULE_FILTERINPUT_AUTHOR_SLUG],
         );
     }
 
@@ -28,6 +30,7 @@ class FilterInputModuleProcessor extends AbstractFormInputModuleProcessor implem
     {
         $filterInputs = [
             self::MODULE_FILTERINPUT_AUTHOR_IDS => [FilterInputProcessor::class, FilterInputProcessor::FILTERINPUT_AUTHOR_IDS],
+            self::MODULE_FILTERINPUT_AUTHOR_SLUG => [FilterInputProcessor::class, FilterInputProcessor::FILTERINPUT_AUTHOR_SLUG],
         ];
         return $filterInputs[$module[1]] ?? null;
     }
@@ -36,6 +39,7 @@ class FilterInputModuleProcessor extends AbstractFormInputModuleProcessor implem
     {
         return match ($module[1]) {
             self::MODULE_FILTERINPUT_AUTHOR_IDS => 'authorIDs',
+            self::MODULE_FILTERINPUT_AUTHOR_SLUG => 'authorSlug',
             default => parent::getName($module),
         };
     }
@@ -44,6 +48,7 @@ class FilterInputModuleProcessor extends AbstractFormInputModuleProcessor implem
     {
         return match ($module[1]) {
             self::MODULE_FILTERINPUT_AUTHOR_IDS => SchemaDefinition::TYPE_ID,
+            self::MODULE_FILTERINPUT_AUTHOR_SLUG => SchemaDefinition::TYPE_STRING,
             default => $this->getDefaultSchemaFilterInputType(),
         };
     }
@@ -68,6 +73,7 @@ class FilterInputModuleProcessor extends AbstractFormInputModuleProcessor implem
     {
         $descriptions = [
             self::MODULE_FILTERINPUT_AUTHOR_IDS => $this->translationAPI->__('Search custom posts from the authors with given IDs', 'pop-users'),
+            self::MODULE_FILTERINPUT_AUTHOR_SLUG => $this->translationAPI->__('Search custom posts from the authors with given slug', 'pop-users'),
         ];
         return $descriptions[$module[1]] ?? null;
     }